requirements.txt
python-telegram-bot==20.3
APScheduler==3.10.4
httpx==0.27.0
Flask==3.0.3
pytz==2024.1
